Have bus interfaces track their neighbor chunks, if necessary.
Needed to properly react to block devices in neighbor chunk being unloaded/loaded. Since we detect devices independent of bus scan, to ensure we can detect removed devices without a bus scan, we need to track this here, too. We cannot rely on the bus controller telling us to re-scan devices once the bus is complete.